refactor: replace boost_filesystem_path_validation with program_options_path for better path handling
