+++ src/tools/contract.cpp
@@ -6,7 +6,7 @@
 #include "util/timezones.hpp"
 #include "util/version.hpp"
 
-#include "util/boost_filesystem_path_validation.hpp"
+#include "util/program_options_path.hpp"
 #include <boost/program_options.hpp>
 #include <boost/program_options/errors.hpp>
