+++ src/tools/customize.cpp
@@ -5,7 +5,7 @@
 #include "util/meminfo.hpp"
 #include "util/version.hpp"
 
-#include "util/boost_filesystem_path_validation.hpp"
+#include "util/program_options_path.hpp"
 #include <boost/program_options.hpp>
 
 #include <filesystem>
