+++ src/tools/store.cpp
@@ -10,7 +10,7 @@
 #include "util/typedefs.hpp"
 #include "util/version.hpp"
 
-#include "util/boost_filesystem_path_validation.hpp"
+#include "util/program_options_path.hpp"
 #include <boost/program_options.hpp>
 
 #include <csignal>
